# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Clinical Review Copilot API (P-194)",
    "description": "Backend API for Clinical Review Copilot platform — FHIR R4 & PDF/OCR Ingestion",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/patients/{patient_id}/verification-items": {
      "get": {
        "tags": [
          "ocr"
        ],
        "summary": "Get Verification Items",
        "operationId": "get_verification_items_api_v1_patients__patient_id__verification_items_get",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Patient Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/verification-items/{verification_item_id}": {
      "patch": {
        "tags": [
          "ocr"
        ],
        "summary": "Patch Verification Item",
        "operationId": "patch_verification_item_api_v1_verification_items__verification_item_id__patch",
        "parameters": [
          {
            "name": "verification_item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Verification Item Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchVerificationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationItem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{document_id}/pages/{page_number}": {
      "get": {
        "tags": [
          "ocr"
        ],
        "summary": "Get Document Page",
        "operationId": "get_document_page_api_v1_documents__document_id__pages__page_number__get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          },
          {
            "name": "page_number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Page Number"
            }
          },
          {
            "name": "representation",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "image",
              "title": "Representation"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{document_id}/raw": {
      "get": {
        "tags": [
          "ocr"
        ],
        "summary": "Get Document Raw",
        "description": "Serve the raw PDF file for in-browser viewing.\n\nFirst checks the in-memory IngestionService DocumentStore.\nFalls back to demo files on disk matching the document_id pattern.",
        "operationId": "get_document_raw_api_v1_documents__document_id__raw_get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PatchVerificationRequest": {
        "properties": {
          "decision": {
            "type": "string",
            "title": "Decision"
          },
          "corrected_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Corrected Text"
          },
          "expected_version": {
            "type": "integer",
            "title": "Expected Version",
            "default": 1
          }
        },
        "type": "object",
        "required": [
          "decision"
        ],
        "title": "PatchVerificationRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VerificationItem": {
        "properties": {
          "verification_item_id": {
            "type": "string",
            "title": "Verification Item Id"
          },
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "page_number": {
            "type": "integer",
            "title": "Page Number"
          },
          "block_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Block Id"
          },
          "bbox": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bbox"
          },
          "extracted_text": {
            "type": "string",
            "title": "Extracted Text"
          },
          "corrected_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Corrected Text"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "verified",
              "dismissed"
            ],
            "title": "Status",
            "default": "pending"
          },
          "engine": {
            "type": "string",
            "title": "Engine",
            "default": "paddleocr"
          },
          "engine_version": {
            "type": "string",
            "title": "Engine Version",
            "default": "3.0.0"
          }
        },
        "type": "object",
        "required": [
          "verification_item_id",
          "document_id",
          "page_number",
          "extracted_text",
          "confidence"
        ],
        "title": "VerificationItem"
      },
      "VerificationListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/VerificationItem"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "page",
          "page_size",
          "total"
        ],
        "title": "VerificationListResponse"
      },
...
    }
  }
}